cpufreq: stats: Fix broken stats and panic on reset
It turns out that if the stats driver starts with an invalid index (-1),
it decides to just give up and go home immediately. The problem is,
it loops like that forever and never actually updates to a valid index.
So we got 0s for everything.

Also fixed a dumb panic when resetting stats because it tried to access
array index -1.

Now monitoring apps like FKM or SmartPack can actually see what the CPU is doing.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>