[IMPORTANT] configs: s5e8825: Re-enable CONFIG_EXYNOS_ADV_TRACER
Without this driver in place, the device will lock up on trying to
reboot most of the time, and it cannot be recovered with the hard reset
hook or any keycombo as the PMIC seems to latch into a bad state.

The only way to recover it after that is to unplug the battery and
replug it. Confirmed on Galaxy A25 and A53.

The issue seemingly occurs because the exynos-bcm_dbg driver depends
on the ADV_TRACER framework to send IPC commands to the BCM firmware.
Apparently, these commands are responsible for synchronizing power
domains and transitioning the hardware states during boot and suspend.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>